mlops-task/requirements.txt
pandas
pyyaml